+++ scripts/hooks/btslab_post.sh
@@ -1,0 +1,1 @@
+docker cp btslab/ btslab:/opt/lampp/htdocs