+++ scripts/hooks/damnvulnrest_post.sh
@@ -1,0 +1,1 @@
+cd Damn-Vulnerable-RESTaurant-API-Game && bash ./start_app.sh && cd ..