+++ scripts/hooks/damnvulnrest_pre.sh
@@ -1,0 +1,1 @@
+git clone https://github.com/theowni/Damn-Vulnerable-RESTaurant-API-Game.git && cd Damn-Vulnerable-RESTaurant-API-Game