+++ scripts/hooks/jvl_post.sh
@@ -1,0 +1,1 @@
+cd JavaVulnerableLab && docker-compose up -d && cd ..