+++ scripts/hooks/jvl_pre.sh
@@ -1,0 +1,1 @@
+git clone https://github.com/CSPF-Founder/JavaVulnerableLab.git && cd JavaVulnerableLab