+++ scripts/hooks/mutillidae_pre.sh
@@ -1,0 +1,1 @@
+git clone https://github.com/webpwnized/mutillidae-docker.git && cd mutillidae-docker && sed -i 's