+++ scripts/hooks/nosqli_post.sh
@@ -1,0 +1,1 @@
+cd nosqlilab && docker-compose up -d --build && cd ..