+++ scripts/hooks/nosqli_pre.sh
@@ -1,0 +1,1 @@
+git clone https://github.com/madamantis-leviathan/nosqlilab.git && cd nosqlilab && sed -i 's/^version: .*$/version: "3.7-3.9"/' docker-compose.yml && sed -i "/web:/at    container_name: nosqli-web" docker-compose.yml && sed -i "s/^t//" docker-compose.yml