+++ scripts/hooks/ssrflab_post.sh
@@ -1,0 +1,1 @@
+cd SSRF-LAB && docker cp ./index.php ssrflab:/var/www/html && cd ..