+++ scripts/hooks/vulnado_post.sh
@@ -1,0 +1,1 @@
+cd vulnado && sudo docker compose up -d && cd ..