+++ scripts/hooks/vulnado_pre.sh
@@ -1,0 +1,1 @@
+git clone https://github.com/ScaleSec/vulnado && cd vulnado